% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-cookies.R
\name{mod_cookies_ui}
\alias{mod_cookies_ui}
\alias{mod_cookies_server}
\title{Cookies shiny module}
\usage{
mod_cookies_ui(id)

mod_cookies_server(input, output, session, name, valid = 30)
}
\arguments{
\item{id}{An id used in the shiny module, required for namespacing
even though no UI elements are created.}

\item{input, output, session}{Arguments to shiny}

\item{name}{The name of the cookie to save/retrieve the username
from.  This should be globally unique and we will set this
cookie if not pesent.}

\item{valid}{The period of validity of the cookie, in days.  We
will refresh the cookie if it is present.}
}
\description{
A shiny module for working with cookies.
}
