% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{mod_remotesave_ui}
\alias{mod_remotesave_ui}
\alias{mod_remotesave_server}
\title{Remote saving shiny module}
\usage{
mod_remotesave_ui(id)

mod_remotesave_server(input, output, session, root, user, get_state,
  set_state, ..., period = 10)
}
\arguments{
\item{id}{An id used to namespace the shiny module}

\item{input, output, session}{Arguments to shiny}

\item{root}{A reactive indicating the root of the remote save (see
\code{\link{remote_save}}}

\item{user}{A reactive value indicating the user name for the
remote save (\code{\link{remote_save}})}

\item{get_state}{A function that takes no arguments and returns
the application state as an arbitrary R object (though suitable
for serialisation, so pointers will not work)}

\item{set_state}{A function that takes one argument, the state as
returned by \code{get_state}, and restores the application
state.  Any return value is ignored.}

\item{...}{Additional arguments passed to
\code{\link{remote_save}}.  Note that these cannot be reactive.}

\item{period}{The period, in seconds, to save.  This is
approximate, and is implemented using
\code{shiny::invalidateLater}.}
}
\description{
A shiny module for saving state into a remote redis server
}
