% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{remote_save}
\alias{remote_save}
\title{Remotely save state}
\usage{
remote_save(root, user, url = NULL, session = NULL)
}
\arguments{
\item{root}{The root within the redis server to save state at}

\item{user}{The username (will be appended to the root)}

\item{url}{The url for the redis server - optional, and using the
\code{REDIS_URL} environment variable may be preferable}

\item{session}{An optional session id as a string.  If missing
then a random string will be used.}
}
\description{
Remotely save state
}
